requirements.txt
scipy==1.3.1
six==1.12.0
numpy==1.17.2
pytest==5.2.1
pandas==0.25.1
tqdm==4.36.1
albumentations==0.4.3
args==0.1.0
jpeg4py==0.1.4
Pillow==7.0.0
scikit_learn==0.22.1
tabulate==0.8.6
tensorboard==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
torch==1.4.0
torchvision==0.5.0
wandb==0.8.27
ipdb==0.13.2
higher==0.2.1
